Z_Tests/OpenAI/requirements.txt
streamlit==1.37.1
#great_expectations==0.18.19
python-dotenv==1.0.1
openai==1.40.0
pandas==2.2.2